Dijkstra pronto

+++ grafo.h
@@ -1,6 +1,8 @@
 #ifndef GRAFO_H
 #define GRAFO_H
 
+
+
 typedef struct mat_grafo Mat_grafo;
 
 Mat_grafo *cria_grafo(int vrt);
@@ -11,5 +13,5 @@ char *le_arquivo(Mat_grafo *mat);
 /*DJK*/
 float **djk_percorre(Mat_grafo *mat, int i_start, int j_start);
 Mat_grafo* cria_visitados(void); //cria lista de visitados
-int foi_visitado(int ** visitados, int i, int j); //verifica se um no foi visitado
+int foi_visitado(float ** visitados, int i, int j); //verifica se um no foi visitado
 #endif
